requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
cssselect2==0.8.0
diff-match-patch==20230430
Django==5.0.7
django-classy-tags==4.1.0
django-filer==3.3.1
django-formtools==2.5.1
django-import-export==4.1.1
django-meta==2.4.2
django-polymorphic==3.1.0
django-sekizai==4.1.0
django-tinymce==4.1.0
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
djangocms-admin-style==3.3.1
djangocms-page-meta==1.4.5
easy-thumbnails==2.10
Faker==28.4.1
idna==3.10
lxml==5.3.1
numpy==2.1.1
pandas==2.2.2
pillow==10.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
reportlab==4.3.1
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
svglib==1.5.1
tablib==3.5.0
tinycss2==1.4.0
typing_extensions==4.13.0
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.3.0
webencodings==0.5.1
